+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Implemented poll publishing.
 - Switch to editing mode when clicking on preview.
+- Expand ellipsized profile fields on click.
 
 ### Changed